refactor: move try_read_to_string fn to setup
- a try_read_to_string function can be shared between kmd and node modules
